+++ SGRH.Api/Controllers/UsuariosController/UsuarioController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using SGRH.Application.DTOs.Usuarios.UsuarioDto;
 using SGRH.Application.Interfaces.Usuarios;
+using System.Threading.Tasks.Dataflow;
 
 namespace SGRH.Api.Controllers.UsuariosController
 {
@@ -34,8 +35,13 @@ namespace SGRH.Api.Controllers.UsuariosController
         [HttpGet("GetById")]
         public async Task<IActionResult> GetById(int Id) 
         {
+            if (Id <= 0)
+                return BadRequest("No se Encontro el Uusario Solicitado");
+
             var result = await _service.GetByIdAsync(Id);
 
+            if (result == null)
+                return BadRequest("No se Encontro el Usuario Solicitado");
 
             return Ok(result);
         }
@@ -43,8 +49,13 @@ namespace SGRH.Api.Controllers.UsuariosController
         [HttpDelete("DeleteUsuario")]
         public async Task<IActionResult> Delete(int Id, int IdUsuario) 
         {
+            if (Id <= 0 || IdUsuario <= 0)
+                return BadRequest("No se Enctontro el Usuario a Eliminar");
+
             var result = await _service.DeleteAsync(Id,IdUsuario);
 
+            if (result == null)
+                return BadRequest("No se Encontro el Uusario a Eliminar");
 
             return Ok(result);
         }
